app/requirements.txt
traci==1.15.0
mysql-connector-python==8.3.0
requests==2.32.0
numpy==2.2.4
scipy==1.15.2